+++ src/Plainion.Windows/Interactivity/MultiSelectBehavior.cs
@@ -4,7 +4,7 @@ using System.Collections.ObjectModel;
 using System.Collections.Specialized;
 using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Interactivity;
+using Microsoft.Xaml.Behaviors;
 
 namespace Plainion.Windows.Interactivity
 {