+++ src/Plainion.Windows/Interactivity/OpenLinkOnCtrlClickBehavior.cs
@@ -3,7 +3,7 @@ using System.Diagnostics;
 using System.Windows.Controls;
 using System.Windows.Documents;
 using System.Windows.Input;
-using System.Windows.Interactivity;
+using Microsoft.Xaml.Behaviors;
 
 namespace Plainion.Windows.Interactivity
 {